requirements.txt
pyautogui
pyperclip
pyobjc-framework-Quartz
requests
